script manifest load change
the server will no longer require the manifest file to be present when started (node will not throw an error if it cannot find the json). it will be loaded later on by manually checking to see if the file exists and then reading it in and parsing the contents to an object. will allow the server to not restart when the file changes and can dynamically update the manifest vs restarting the server like it has been